Refactor e2e codes
