+++ test/e2e/edgesyncer/testdata/edgesyncconfig/edge-sync-config.yaml
