+++ test/e2e/edgesyncer/testdata/edgesyncconfig/sample-cr-downsync.yaml
