+++ test/e2e/edgesyncer/testdata/edgesyncconfig/sample-cr-upsync.yaml
