+++ test/e2e/edgesyncer/testdata/edgesyncconfig/sample-crd.yaml
