+++ test/e2e/edgesyncer/testdata/syncerconfig/sample-downsync-cr.yaml
