+++ test/e2e/edgesyncer/testdata/syncerconfig/sample-downsync-crd.yaml
