+++ test/e2e/edgesyncer/testdata/syncerconfig/sample-upsync-cr.yaml
