+++ test/e2e/edgesyncer/testdata/syncerconfig/syncer-config-wildcard.yaml
