+++ test/e2e/edgesyncer/testdata/syncerconfig/syncer-config.yaml
